labs/lab09-detection-pipeline/starter/main.py
#!/usr/bin/env python3
"""
Lab 09: Multi-Stage Threat Detection Pipeline - Starter Code

Build an end-to-end threat detection pipeline combining ML and LLM components.
"""

import os
import json
import uuid
from typing import List, Dict, Optional
from datetime import datetime
from pathlib import Path

import numpy as np
from sklearn.ensemble import IsolationForest

from dotenv import load_dotenv
load_dotenv()

try:
    from langchain_anthropic import ChatAnthropic
    LANGCHAIN_AVAILABLE = True
except ImportError:
    LANGCHAIN_AVAILABLE = False

from rich.console import Console
console = Console()


# =============================================================================
# Task 1: Data Ingestion Layer
# =============================================================================

class EventIngestor:
    """Ingest and normalize security events."""

    def __init__(self):
        self.buffer = []

    def ingest_event(self, raw_event: dict, source: str) -> dict:
        """
        Normalize a raw event into standard schema.

        TODO:
        1. Parse event based on source type
        2. Extract standard fields
        3. Normalize timestamps
        4. Return normalized event
        """
        # YOUR CODE HERE
        pass

    def create_normalized_event(self, raw: dict, source: str) -> dict:
        """
        Create normalized event structure.

        Standard schema:
        {
            "id": "uuid",
            "timestamp": "ISO8601",
            "source": "sysmon|windows|network",
            "event_type": "process|network|file|auth",
            "host": "hostname",
            "user": "username",
            "details": {...},
            "raw": original_event
        }
        """
        # YOUR CODE HERE
        pass


# =============================================================================
# Task 2: ML Filtering Stage
# =============================================================================

class MLFilterStage:
    """Stage 1: ML-based anomaly filtering."""

    def __init__(self, contamination: float = 0.05):
        self.contamination = contamination
        self.model = None
        self.threshold = 0.7

    def extract_features(self, event: dict) -> np.ndarray:
        """
        Extract ML features from event.

        TODO:
        1. Extract numeric features
        2. Encode categorical features
        3. Return feature vector
        """
        # YOUR CODE HERE
        pass

    def train(self, events: List[dict]):
        """
        Train the anomaly detection model.

        TODO:
        1. Extract features from all events
        2. Train Isolation Forest
        """
        # YOUR CODE HERE
        pass

    def score_event(self, event: dict) -> float:
        """
        Score event anomaly level (0-1).

        TODO:
        1. Extract features
        2. Get anomaly score from model
        3. Normalize to 0-1
        """
        # YOUR CODE HERE
        pass

    def filter_events(self, events: List[dict]) -> List[dict]:
        """
        Filter events above threshold.

        TODO:
        1. Score all events
        2. Keep events > threshold
        3. Add score to event
        """
        # YOUR CODE HERE
        pass


# =============================================================================
# Task 3: LLM Enrichment Stage
# =============================================================================

class LLMEnrichmentStage:
    """Stage 2: LLM-based context enrichment."""

    def __init__(self, llm=None):
        self.llm = llm
        self.cache = {}

    def enrich_event(self, event: dict) -> dict:
        """
        Enrich event with LLM analysis.

        TODO:
        1. Format event for LLM
        2. Get analysis (threat assessment, MITRE mapping)
        3. Parse and add enrichments
        4. Return enriched event
        """
        # YOUR CODE HERE
        pass


# =============================================================================
# Task 4: Correlation Stage
# =============================================================================

class CorrelationStage:
    """Stage 3: Event correlation and chain detection."""

    def __init__(self, time_window: int = 300):
        self.time_window = time_window
        self.event_buffer = []

    def add_event(self, event: dict):
        """Add event to correlation buffer."""
        # YOUR CODE HERE
        pass

    def find_related_events(self, event: dict) -> List[dict]:
        """
        Find events related to this one.

        TODO:
        1. Search buffer for related events
        2. Apply time window filter
        3. Match on host, user, or process
        """
        # YOUR CODE HERE
        pass

    def detect_attack_chain(self, events: List[dict]) -> dict:
        """
        Detect attack chain patterns.

        TODO:
        1. Order events by time
        2. Map to ATT&CK tactics
        3. Look for attack patterns
        """
        # YOUR CODE HERE
        pass


# =============================================================================
# Task 5: Verdict & Response Stage
# =============================================================================

class VerdictStage:
    """Stage 4: Final verdict and response generation."""

    def __init__(self, llm=None):
        self.llm = llm

    def generate_verdict(self, events: List[dict]) -> dict:
        """
        Generate final verdict.

        TODO:
        1. Analyze all evidence
        2. Calculate confidence
        3. Determine verdict
        4. Generate explanation
        """
        # YOUR CODE HERE
        pass

    def create_alert(self, events: List[dict], verdict: dict) -> dict:
        """
        Create final alert for SOC.

        TODO:
        1. Format alert structure
        2. Include evidence
        3. Add response actions
        """
        # YOUR CODE HERE
        pass


# =============================================================================
# Task 6: Pipeline Orchestrator
# =============================================================================

class DetectionPipeline:
    """Orchestrate the complete pipeline."""

    def __init__(self, config: dict = None):
        config = config or {}
        self.ingestor = EventIngestor()
        self.ml_filter = MLFilterStage()
        self.correlator = CorrelationStage()
        # Add other stages

    def process_event(self, raw_event: dict, source: str) -> Optional[dict]:
        """
        Process single event through pipeline.

        TODO:
        1. Normalize event
        2. Run through ML filter
        3. Enrich if suspicious
        4. Correlate
        5. Generate verdict if needed
        """
        # YOUR CODE HERE
        pass

    def process_batch(self, events: List[dict]) -> List[dict]:
        """Process batch of events."""
        # YOUR CODE HERE
        pass


# =============================================================================
# Main
# =============================================================================

def main():
    """Main execution."""
    console.print("[bold]Lab 09: Threat Detection Pipeline[/bold]")

    # Create sample events
    sample_events = [
        {
            "timestamp": "2024-01-15T03:22:10Z",
            "host": "WORKSTATION01",
            "event_type": "process",
            "process_name": "powershell.exe",
            "command_line": "powershell -enc SGVsbG8gV29ybGQ=",
            "parent_process": "cmd.exe",
            "user": "jsmith"
        },
        {
            "timestamp": "2024-01-15T03:22:15Z",
            "host": "WORKSTATION01",
            "event_type": "network",
            "process_name": "powershell.exe",
            "dest_ip": "185.143.223.47",
            "dest_port": 443,
            "user": "jsmith"
        }
    ]

    console.print(f"\n[yellow]Processing {len(sample_events)} sample events...[/yellow]")

    pipeline = DetectionPipeline()

    for event in sample_events:
        result = pipeline.process_event(event, "sysmon")
        if result:
            console.print(f"[green]Alert generated![/green]")
        else:
            console.print("No alerts (complete the TODO sections)")

    console.print("\nComplete the TODO sections to enable detection!")


if __name__ == "__main__":
    main()
