requirements.txt
# AI for the Win - Dependencies
# Security AI Training Program

# Core ML/AI Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Deep Learning (optional, for autoencoders)
torch>=2.0.0,<3.0.0

# LLM Frameworks
langchain>=0.1.0,<1.0.0
langchain-anthropic>=0.1.0,<1.0.0
langchain-community>=0.0.20,<1.0.0
langgraph>=0.0.30,<1.0.0

# Vector Database
chromadb>=0.4.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0

# NLP
nltk>=3.8.0,<4.0.0
tiktoken>=0.5.0,<1.0.0

# Security Tools
yara-python>=4.3.0,<5.0.0
pefile>=2023.2.7

# API & Web
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.9.0,<4.0.0

# Document Processing
pypdf>=3.17.0,<5.0.0
docx2txt>=0.8

# Web Framework (for APIs and UIs)
fastapi>=0.104.0,<1.0.0
uvicorn>=0.24.0,<1.0.0
streamlit>=1.28.0,<2.0.0

# CLI & Visualization
rich>=13.6.0,<14.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

# Data Validation
pydantic>=2.5.0,<3.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: Network Analysis
# python-nmap>=0.7.1

# Optional: Caching/Queuing
# redis>=5.0.0
